Queries to separate python file
